refactor: 투표 결과 조회 비즈니스 로직을 서비스로 이전 컨트롤러의 winner 집계 로직을 VoteService.getVoteResults()로 이동하고 categoryService 의존성도 서비스로 이전
